+++ src/heat.js
@@ -72,6 +72,7 @@
                     bindingOptions = buildAttributeOptions( bindingOptions.result );
                     bindingOptions.element = element;
                     bindingOptions.currentView = {};
+                    bindingOptions.currentView.colorsVisible = {};
 
                     fireCustomTrigger( bindingOptions.onBeforeRender, element );
 
@@ -269,20 +270,20 @@
         };
 
         var mapRangeColorsLength = bindingOptions.mapRangeColors.length,
-            className = null;
+            useMapRangeColor = null;
     
         for ( var mapRangeColorsIndex = 0; mapRangeColorsIndex < mapRangeColorsLength; mapRangeColorsIndex++ ) {
             var mapRangeColor = mapRangeColors[ mapRangeColorsIndex ];
 
             if ( dateCount >= mapRangeColor.minimum ) {
-                className = _string.space + mapRangeColor.cssClassName;
+                useMapRangeColor = mapRangeColor;
             } else {
                 break;
             }
         }
 
-        if ( isDefined( className ) ) {
-            day.className += _string.space + className;
+        if ( isDefined( useMapRangeColor ) && isHeatMapColorVisible( bindingOptions, useMapRangeColor.id ) ) {
+            day.className += _string.space + useMapRangeColor.cssClassName;
         }
 
         if ( !isDefined( _elements_Day_Width ) && !bindingOptions.showMonthDayGaps ) {
@@ -312,16 +313,45 @@
             var mapRangeColorsLength = mapRangeColors.length;
     
             for ( var mapRangeColorsIndex = 0; mapRangeColorsIndex < mapRangeColorsLength; mapRangeColorsIndex++ ) {
-                var mapRangeColor = mapRangeColors[ mapRangeColorsIndex ];
-    
-                var day = createElement( "div", "day " + mapRangeColor.cssClassName );
-                days.appendChild( day );
+                renderControlViewGuideDay( bindingOptions, days, mapRangeColors[ mapRangeColorsIndex ] );
             }
     
             var moreText = createElement( "div", "more-text" );
             moreText.innerHTML = _configuration.moreText;
             guide.appendChild( moreText );
         }
+    }
+
+    function renderControlViewGuideDay( bindingOptions, days, mapRangeColor ) {
+        var day = createElement( "div" );
+        day.title = mapRangeColor.tooltipText;
+        days.appendChild( day );
+
+        if ( isHeatMapColorVisible( bindingOptions, mapRangeColor.id ) ) {
+            day.className = "day " + mapRangeColor.cssClassName;
+        } else {
+            day.className = "day";
+        }
+
+        day.onclick = function() {
+            if ( !bindingOptions.currentView.colorsVisible.hasOwnProperty( mapRangeColor.id ) ) {
+                bindingOptions.currentView.colorsVisible[ mapRangeColor.id ] = true;
+            }
+
+            if ( bindingOptions.currentView.colorsVisible[ mapRangeColor.id ] ) {
+                day.className = "day";
+            } else {
+                day.className = "day " + mapRangeColor.cssClassName;
+            }
+
+            bindingOptions.currentView.colorsVisible[ mapRangeColor.id ] = !bindingOptions.currentView.colorsVisible[ mapRangeColor.id ];
+
+            renderControl( bindingOptions );
+        };
+    }
+
+    function isHeatMapColorVisible( bindingOptions, id ) {
+        return !bindingOptions.currentView.colorsVisible.hasOwnProperty( id ) || bindingOptions.currentView.colorsVisible[ id ];
     }
 
 
@@ -351,21 +381,33 @@
         options.mapRangeColors = getDefaultArray( options.mapRangeColors, [
             {
                 minimum: 10,
-                cssClassName: "day-color-1"
+                cssClassName: "day-color-1",
+                tooltipText: "Day Color 1"
             },
             {
                 minimum: 15,
-                cssClassName: "day-color-2"
+                cssClassName: "day-color-2",
+                tooltipText: "Day Color 2"
             },
             {
                 minimum: 20,
-                cssClassName: "day-color-3"
+                cssClassName: "day-color-3",
+                tooltipText: "Day Color 3"
             },
             {
                 minimum: 25,
-                cssClassName: "day-color-4"
+                cssClassName: "day-color-4",
+                tooltipText: "Day Color 4"
             }
         ] );
+
+        var mapRangeColorsLength = options.mapRangeColors.length;
+
+        for ( var mapRangeColorsIndex = 0; mapRangeColorsIndex < mapRangeColorsLength; mapRangeColorsIndex++ ) {
+            if ( !isDefinedString( options.mapRangeColors[ mapRangeColorsIndex ].id ) ) {
+                options.mapRangeColors[ mapRangeColorsIndex ].id = newGuid();
+            }
+        }
 
         options = buildAttributeOptionStrings( options );